separate file for tests
and added comments
