+++ index.js
@@ -1,18 +1,31 @@
+const { existsSync, writeFileSync } = require('node:fs');
 const fs = require('node:fs');
+const { builtinModules } = require('node:module');
 
 
 // Definition de DB 
 class DB {
  
-  // ... Retourner tous les messages storees
+  
+  /**
+   * get all messages in DB
+   * @returns obj message list
+   */
   static getMessages() {
-
-    let words = JSON.parse(fs.readFileSync('./messages.json','utf-8'));
-    return words;
+    if(fs.existsSync('./messages.json')){
+      let words = JSON.parse(fs.readFileSync('./messages.json','utf-8'));
+      return words;
+    }
+    else{
+      fs.writeFileSync('./messages.json','[]');
+      return [];
+    }
   }
 
-
-  // ... prend un objet `msg` et l'ajoute a `./messages.json`
+  /**
+   * add a message to DB
+   * @param {Message} msg - the message to add to the DB 
+   */
   static addMessage(msg) {
 
     let msgs = DB.getMessages();
@@ -23,17 +36,6 @@ class DB {
 
 }
 
-// Utilisation de DB
-
-// avoir une liste de tous les objets `msg`
-const allMessages = DB.getMessages();
-console.log(allMessages)
-
-// Ajouter un message a notre base de donnees
-DB.addMessage({ 
-  user: 'myName',
-  msg: 'this is a sample message'
-});
-
+module.exports = DB;
