+++ messages.json
@@ -1,0 +1,1 @@
+[{"user":"myName","msg":"this is a sample message"},{"user":"myName","msg":"this is a sample message"}]
